+++ templates/pages/dashboard.html
@@ -6,15 +6,15 @@
     <h2><strong>Total allocated: {{ .Budget.Allocated }} {{ .Budget.Currency }}</strong></h2>
     <h3>Remaining budget: {{ .BudgetRemaining }}</h3>
     <hr>
-    <!-- <h2>Categories</h2> -->
-    <!-- {{ range .Budget.BudgetCategories }} -->
-        <!-- <hr> -->
-        <!-- <p style="background-color: {{ .Category.Color }}"><strong>{{ .Category.Name }}</strong></p> -->  
-        <!-- <p><em>{{ .Category.Description }}</em></p> -->
-        <!-- <h3>Allocated: {{ .Allocated }}</h3> -->
-        <!-- <h3>Remaining: {{ .Remaining }}</h3> -->
-        <!-- <hr> -->
-    <!-- {{end}} -->
+    <h2>Categories</h2>
+    {{ range .Budget.BudgetCategories }}
+        <hr>
+        <p style="background-color: {{ .Category.Color }}"><strong>{{ .Category.Name }}</strong></p>  
+        <p><em>{{ .Category.Description }}</em></p>
+        <h3>Allocated: {{ .Allocated }}</h3>
+        <h3>Remaining: {{ .Remaining }}</h3>
+        <hr>
+    {{end}}
 
     <h2>Transaction Log</h2>
     {{ range .Budget.BudgetCategories }}
